fix(all): lock siswa in ekstrakurikuler

+++ apps/server/src/guru/ekstrakurikuler/ekstrakurikuler.service.ts
@@ -2,6 +2,7 @@ import { TRPCError } from '@trpc/server';
 import { PrismaService } from '../../prisma/prisma.service';
 import { $Enums } from '@prisma/client';
 import { Injectable } from '@nestjs/common';
+import { isRaportLocked, isSubset, setDifference } from '../../utils';
 
 export interface NilaiEkstrakurikuler {
   id_siswa: string;
@@ -62,6 +63,7 @@ export class GuruEkstrakurikulerService {
   }
 
   async getAnggotaList(sessionUsername: string, id: string) {
+    const ekstrakurikuler = await this.ensureCanEdit(sessionUsername, id);
     const result = await this.prismaClient.ekstrakurikuler.findUnique({
       where: {
         id_esktrakurikuler: id,
@@ -78,21 +80,29 @@ export class GuruEkstrakurikulerService {
                 nama: true,
                 NIS: true,
                 NISN: true,
+                Raport: {
+                  where: {
+                    id_periode_ajar: ekstrakurikuler.id_periode_ajar,
+                  },
+                  take: 1,
+                  select: {
+                    status: true,
+                  },
+                },
               },
             },
           },
         },
       },
     });
-
     if (!result) this.throwNotFound();
-    if (result.username_guru != sessionUsername) this.throwForbidden();
 
     return result.Anggota_Ekstrakurikuler.map(
-      ({ nilai, keterangan, Siswa }) => ({
+      ({ nilai, keterangan, Siswa: { Raport, ...Siswa } }) => ({
         ...Siswa,
         nilai,
         keterangan,
+        is_locked: isRaportLocked(Raport.at(0)?.status),
       })
     );
   }
@@ -104,11 +114,14 @@ export class GuruEkstrakurikulerService {
       },
       select: {
         username_guru: true,
+        id_periode_ajar: true,
       },
     });
 
     if (!data) this.throwNotFound();
     if (data.username_guru != sessionUsername) this.throwForbidden();
+
+    return data;
   }
 
   async updateAnggotaList(
@@ -116,7 +129,60 @@ export class GuruEkstrakurikulerService {
     id: string,
     siswaIdList: string[]
   ) {
-    await this.ensureCanEdit(sessionUsername, id);
+    const ekstrakurikuler = await this.ensureCanEdit(sessionUsername, id);
+
+    const lockedAnggota = new Set(
+      (
+        await this.prismaClient.anggota_Ekstrakurikuler.findMany({
+          where: {
+            id_ekstrakurikuler: id,
+            Siswa: {
+              Raport: {
+                some: {
+                  id_periode_ajar: ekstrakurikuler.id_periode_ajar,
+                  status: {
+                    not: 'MENUNGGU_KONFIRMASI',
+                  },
+                },
+              },
+            },
+          },
+          select: {
+            id_siswa: true,
+          },
+        })
+      ).map((Anggota) => Anggota.id_siswa)
+    );
+    const currentAnggota = new Set(siswaIdList);
+
+    if (!isSubset(lockedAnggota, currentAnggota))
+      throw new TRPCError({
+        code: 'FORBIDDEN',
+        message: 'Cannot delete locked anggota',
+      });
+
+    const otherAnggota = setDifference(currentAnggota, lockedAnggota);
+
+    const newLockedAnggota = await this.prismaClient.raport.findMany({
+      where: {
+        id_periode_ajar: ekstrakurikuler.id_periode_ajar,
+        id_siswa: {
+          in: [...otherAnggota],
+        },
+        status: {
+          not: 'MENUNGGU_KONFIRMASI',
+        },
+      },
+      select: {
+        id_siswa: true,
+      },
+    });
+
+    if (newLockedAnggota.length > 0)
+      throw new TRPCError({
+        code: 'FORBIDDEN',
+        message: 'Cannot add locked anggota',
+      });
 
     await this.prismaClient.$transaction([
       this.prismaClient.anggota_Ekstrakurikuler.deleteMany({
@@ -142,7 +208,7 @@ export class GuruEkstrakurikulerService {
     id: string,
     nilai: NilaiEkstrakurikuler[]
   ) {
-    await this.ensureCanEdit(sessionUsername, id);
+    const ekstrakurikuler = await this.ensureCanEdit(sessionUsername, id);
 
     const validList = await this.prismaClient.anggota_Ekstrakurikuler.findMany({
       where: {
@@ -150,17 +216,37 @@ export class GuruEkstrakurikulerService {
         id_siswa: {
           in: nilai.map((nilai) => nilai.id_siswa),
         },
+        Siswa: {
+          OR: [
+            {
+              Raport: {
+                none: {
+                  id_periode_ajar: ekstrakurikuler.id_periode_ajar,
+                },
+              },
+            },
+            {
+              Raport: {
+                some: {
+                  id_periode_ajar: ekstrakurikuler.id_periode_ajar,
+                  status: "MENUNGGU_KONFIRMASI"
+                },
+              },
+            },
+          ],
+        },
       },
       select: {
         id_siswa: true,
       },
     });
 
-    if (validList.length != nilai.length)
+    if (validList.length != nilai.length) {
       throw new TRPCError({
         code: 'BAD_REQUEST',
         message: 'Invalid siswa found',
       });
+    }
 
     await this.prismaClient.$transaction(
       nilai.map(({ id_siswa, nilai, keterangan }) =>
